Add serlogin program, version 0.4.0
This has support for logging in, with device on other end of
serial port being at a shell prompt (either needing to switch user
or not) or at a login prompt.  It works for accounts with and without
passwords on a beaglebone running a Debian distribution.

Testing on more platforms with other distros is welcome.